test: move the six largest inline test modules to sibling files
